+++ tests/Spec/Data/Type/EqualitySpec.hs
@@ -13,6 +13,9 @@ Portability: GHC
 -}
 module Spec.Data.Type.EqualitySpec (main, spec) where
 
+import Data.Proxy.Compat (Proxy(..))
+import Data.Type.Equality.Compat
+
 import Instances.Data.Type.Equality ()
 
 import Prelude ()
@@ -20,29 +23,17 @@ import Prelude.Compat
 
 import Test.Hspec (Spec, hspec, parallel)
 
-#if MIN_VERSION_base(4,7,0)
-import Data.Proxy.Compat (Proxy(..))
-import Data.Type.Equality ((:~:))
-# if MIN_VERSION_base(4,10,0)
-import Data.Type.Equality ((:~~:))
-# endif
-
 import Spec.Utils (matchesTextShowSpec)
 import Test.Hspec (describe)
-#endif
 
 main :: IO ()
 main = hspec spec
 
 spec :: Spec
 spec = parallel $ do
-#if MIN_VERSION_base(4,7,0)
     describe "Int :~: Int" $
         matchesTextShowSpec (Proxy :: Proxy (Int :~: Int))
-# if MIN_VERSION_base(4,10,0)
+# if MIN_VERSION_base(4,9,0)
     describe "Int :~~: Int" $
         matchesTextShowSpec (Proxy :: Proxy (Int :~~: Int))
 # endif
-#else
-    pure ()
-#endif